関連モデル(リレーション)レコード削除の設定

+++ src/mealbox/app/Http/Controllers/Admin/AdminController.php
@@ -138,6 +138,10 @@ class AdminController extends Controller
      */
     function adminOrderDelete($id)
     {
+        if (empty($id)) {
+            return redirect()->route('admin.order_list');
+        }
+
         try {
             Order::destroy($id);
         } catch(\Throwable $e) {
